app/requirements.txt
snowflake-connector-python
pandas
Flask